ShipDataset.py
from torch.utils.data import Dataset
from torchvision import transforms
from PIL import Image
from until import *
import os

transform_map = {'train': transforms.Compose([
    transforms.RandomResizedCrop(224),
    transforms.RandomHorizontalFlip(),
    transforms.ToTensor(),
    transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
]), 'val': transforms.Compose([
    transforms.Resize(256),
    transforms.CenterCrop(224),
    transforms.ToTensor(),
    transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
])}

class ShipDataset(Dataset):

    def __init__(self, root, txt, transform=None, mode='train'):
        self.img_path = []
        self.labels = []
        self.transform = transform_map[mode]
        with open(txt) as f:
            for line in f:
                self.img_path.append(os.path.join(root, line.split()[0]))
                self.labels.append(int(line.split()[1]))

    def __len__(self):
        return len(self.labels)

    def __getitem__(self, index):

        path = self.img_path[index]
        label = self.labels[index]
        with open(path, 'rb') as f:
            sample = Image.open(f).convert('RGB')

        if self.transform is not None: # 数据预处理
            sample = self.transform(sample)

        return sample, label, index